+++ views/reports.list.php
@@ -7,6 +7,56 @@ use Modules\MoreReporting\Includes\ReportTypeRegistry;
  * @var array $data
  */
 
+$filter = (new CFilter())
+	->addVar('action', 'morereporting.list')
+	->setResetUrl((new CUrl('zabbix.php'))->setArgument('action', 'morereporting.list'))
+	->setProfile('web.morereporting.list.filter')
+	->setActiveTab($data['active_tab'])
+	->addFilterTab(_('Filter'), [
+		(new CFormGrid())
+			->addClass(CFormGrid::ZBX_STYLE_FORM_GRID_LABEL_WIDTH_TRUE)
+			->addItem([
+				new CLabel(_('Name'), 'filter_name'),
+				new CFormField(
+					(new CTextBox('filter_name', $data['filter']['name']))
+						->setWidth(ZBX_TEXTAREA_FILTER_STANDARD_WIDTH)
+						->setAttribute('autofocus', 'autofocus')
+				)
+			])
+			->addItem([
+				new CLabel(_('Type')),
+				new CFormField(
+					(new CSelect('filter_report_type'))
+						->setValue($data['filter']['report_type'])
+						->addOption(new CSelectOption('', _('Any')))
+						->addOptions(CSelect::createOptionsFromArray($data['report_type_labels']))
+				)
+			]),
+		(new CFormGrid())
+			->addClass(CFormGrid::ZBX_STYLE_FORM_GRID_LABEL_WIDTH_TRUE)
+			->addItem([
+				new CLabel(_('Status')),
+				new CFormField(
+					(new CRadioButtonList('filter_status', $data['filter']['status'] === ''
+						? -1 : (int) $data['filter']['status']
+					))
+						->addValue(_('Any'), -1)
+						->addValue(_('Enabled'), ZBX_REPORT_STATUS_ENABLED)
+						->addValue(_('Disabled'), ZBX_REPORT_STATUS_DISABLED)
+						->setModern(true)
+				)
+			])
+			->addItem([
+				new CLabel(_('Show')),
+				new CFormField(
+					(new CRadioButtonList('filter_show', $data['filter']['show']))
+						->addValue(_('All'), ZBX_REPORT_FILTER_SHOW_ALL)
+						->addValue(_('Created by me'), ZBX_REPORT_FILTER_SHOW_MY)
+						->setModern(true)
+				)
+			])
+	]);
+
 $html_page = (new CHtmlPage())
 	->setTitle($data['title'])
 	->setControls(
@@ -17,9 +67,10 @@ $html_page = (new CHtmlPage())
 				)
 			)
 		))->setAttribute('aria-label', _('Content controls'))
-	);
+	)
+	->addItem($filter);
 
-$table = (new CTableInfo())->setHeader([_('Name'), _('Type'), _('Actions')]);
+$table = (new CTableInfo())->setHeader([_('Name'), _('Type'), _('Status'), _('Created by'), _('Actions')]);
 
 foreach ($data['reports'] as $report) {
 	$run_action = ReportTypeRegistry::action($report['report_type']);
@@ -31,6 +82,20 @@ foreach ($data['reports'] as $report) {
 	$edit_url = (new CUrl('zabbix.php'))
 		->setArgument('action', 'morereporting.report.edit')
 		->setArgument('reportid', $report['reportid']);
+
+	$is_enabled = (int) $report['status'] === ZBX_REPORT_STATUS_ENABLED;
+
+	$status_form = (new CForm('post'))
+		->cleanItems()
+		->setAction((new CUrl('zabbix.php'))->setArgument('action', 'morereporting.report.status')->getUrl())
+		->addVar('_csrf_token', CCsrfTokenHelper::get('morereporting.report.status'))
+		->addVar('reportid', $report['reportid'])
+		->addVar('status', $is_enabled ? ZBX_REPORT_STATUS_DISABLED : ZBX_REPORT_STATUS_ENABLED)
+		->addItem(
+			(new CSubmitButton($is_enabled ? _('Enabled') : _('Disabled')))
+				->addClass(ZBX_STYLE_BTN_LINK)
+				->addClass($is_enabled ? ZBX_STYLE_GREEN : ZBX_STYLE_GREY)
+		);
 
 	$delete_form = (new CForm('post'))
 		->cleanItems()
@@ -45,9 +110,13 @@ foreach ($data['reports'] as $report) {
 				)
 		);
 
+	$creator = $data['users'][$report['userid']] ?? null;
+
 	$table->addRow([
 		new CLink($report['name'], $run_url->getUrl()),
 		$data['report_type_labels'][$report['report_type']] ?? $report['report_type'],
+		$status_form,
+		$creator !== null ? getUserFullname($creator) : '',
 		[
 			new CLink(_('Edit'), $edit_url->getUrl()),
 			' ',
@@ -56,6 +125,6 @@ foreach ($data['reports'] as $report) {
 	]);
 }
 
-$html_page->addItem($table);
-
-$html_page->show();
+$html_page
+	->addItem($table)
+	->show();